Move domElements to utils
