+++ src/constructors/styled.js
@@ -1,7 +1,7 @@
 // @flow
 import css from './css'
 import type { Interpolation, Target } from '../types'
-import domElements from './domElements'
+import domElements from '../utils/domElements'
 
 export default (styledComponent: Function) => {
   const styled = (tag: Target) =>
