+++ src/utils/domElements.js
@@ -134,6 +134,6 @@ export default [
   'stop',
   'svg',
   'text',
-  'tspan'
+  'tspan',
 ]
